E-Series: Resolve issue with delete volume
When the host iqn was not properly defined on a host, it was possible
to define a new volume using docker volume create, but not possible
to delete the volume.

This patch changes an error into a warning message to allow the
deletion to proceed unless the volume has been mapped to a different
host.

resolves #22