Fix issue of '_' not being valid for SF vol names
This patch actually cleans up two things:
1. SolidFire doesn't allow '_' characters in volume names
  This is a problem because things like Compose can preface
  a volume name with 'project_'

  Rather than try and map name changes, we'll instead just stuff the
  docker volume name into a volume attribute and use that as our handle
  going forward.  That means ANY docker valid name is now a valid name
  for a SolidFire Volume.

2. While working on this things got a bit wonky with the prefix
  NDVP by default uses a prefix for Volume and Snapshot names.  Some
  times this is good, some times it's not.  Rather than try and always
  keep that sort of thing tracked, we again just use a *pure* docker
  volume name in the attributes for things.  So no matter what when we
  use the docker volume API and it's representation of the volume name
  things just work.

The whole business of trying to tracking via names by adding prefixes has
a number of issues, not only on the backends themeselves due to character
restrictions as pointed out here, but we also introduce a potential for
corruption in cases like a multi-node deployment where the prefix doesn't
match.

We can at least mitigate some issue by just using the raw docker volume
name as an attribute and use that as a handle as opposed to trying to track and
modify names.

One other concern that should be considered here is the max number of chars for
each backend on the name, but that's another issue for another time.

Addresses Github Issue: #32