Fix up name prefix messiness
One of the things added to the ndvp driver in it's initial release was a
name-prefix default that would prepend all resource names with a default
of `netappdvp-`.  We would perform all look ups using the volume name and
this prefix.

Due to some incompatabilities between products and the use of `-` vs `_`,
we modified the SolidFire driver to use attributes as the mapping method
instead of the volume-name and juggling the prefix. This broke access to
legacy installs that upgraded, becuase they wouldn't have the attribute
and thus never find the volumes.

The name prefixes are actually just troublesome, between upgrade issues like
the one described, and errors in deployments (other ndvp nodes not matching
their configs) it's just better to not use prefixes.

This change does away with prefixes for the SolidFire driver by checking at the
main ndvp interface and NOT munging the name.  We then leave it up to the
driver to do any name munging if neeeded.

To address the issue of legacy volumes after an upgrade, I implemented a
wrapper method to getVolumes that will use the old method of looking by name
with prefix if the search by attributes fails.