Fix SF Naming issues
This patch fixes up naming issues that have been plaguing the
SolidFire backend option in NDVP.

In order to fix this we do a number of things, like skipping the
use of name prefixes, and just using docker-names (we can still
identify telemetry data via volume attributes).

This means we need to do some extra effort with detecting legacy
names, as well as transposed '_' to '-' symbols in names, so it
requires a bit more effort on the get and list operations.

We also fix things up in terms of writing attributes to clones, and
we start removing the reliance on the file-system mount points as the
source of truth for volume existence for SolidFire by moving that into
a listForAccount API call.

Note that if you share your docker account ID for other things, then
you're doing it wrong, stop doing that. :)

This is a temporary quick fix solution, future changes will be proposed
that break these pieces out a bit and simplify the config as well as the
operations.  This is the first step in that direction, eventually we'll
ease testing and maintenance by starting this sort of effort.