Add Timestamp to the non-debug logger
Logrus is cool, but the defaults are sort of annoying, things like no tty
having a different format, and Debug having Timestamps by default but higher
levels not.

This patch just adds Timestamp explicitly so it should be listed in every
message regardless of log level etc.  There's lots of room for improvement in
our logging but this is at least one samll step forward for reading gophers.