Retry in SF on Clone SliceNotRegistered Errors
When using automation, especially in test enviornments it's common
to create a volume and immediately attempt to Clone it.  This is
sometimes problematic for the SolidFire cluster becasue teh Slice
ID's may not be registered yet even though the volume is available
and listed as Active.

We can deal with this pretty easily by just detecting the issue and
putting a retry loop around the Clone request and give the volume
several seconds to settle out.

Addresses Issue: #DE299