Reduce some logging noise in sfapi
I thought it would be nice to have the json request and response
logged as debug in the ndvp logs.  Turns out that it wasn't terribly
useful, and while PrettyPrinting the Json looked nice to a human
it pretty much sucked for data gathering tools.

This change modifies the logging of the request being sent to the
SF API from a multi-line PrettyPrint output in the log file, to
a single line string.

This change also removes the dump of the json response on every call,
we still catch errors and decode, log them, but we get rid of the whole
logging several hundred lines of response for a single call like List.

Might want to put some of this back in the future as options, but would
probably be good to dump them to a second file like ClusterAPI.log or
some clever name.