Retry in SF on Clone SliceNotRegistered Errors
When using automation, especially in test environments it's common
to create a volume and immediately attempt to Clone it. This is
sometimes problematic for the SolidFire cluster because the Slice
ID's may not be registered yet even though the volume is available
and listed as Active.

We can deal with this pretty easily by just detecting the issue and
putting a retry loop around the Clone request and give the volume
several seconds to settle out.

This takes a number of things that I had in flight and starts cleaning
them up as well as breaking them down into smaller more focused
commits.

Addresses Issue: #DE299