Improve Get/List Effeciencies for SolidFire
We got a little in the weeds with the List/Get routines when
we started mucking with Names, Prefixes and Attributes.  The
result I noticed is we were doing something on the order of
4 or 5 List calls to the SF Cluster for each Get.  That's
super lame.

We can fix this up by getting rid of the silly api.volume functions
for each case that I introduced, and instead us a single standard
ListAccountVolumes and process the data rather than get fresh data
each time.

This not only reduces the calls and noise out to the Cluster, but
it also makes things much simpler to debug and track, the search
and filter is now in a single method and is super simple.