Make solidfire_san.getVolume public
Moving consumer specific logic out of sfapi, but there is some coupling
that exists between Trident and NDVP that used the GetVolumeByName helper
that was in the sfapi.

That same functionality is available in the NDVP storage_driver pkg, which
Trident has access to.  Let's just make the helper method Publicly accessible
and then we can use it directly from Trident.