Add VagIDs option to SolidFire
Trident utilizes SolidFire VAGs for iSCSI connection access control.
This works great, but currently things are setup for a single default
VAG, and each VAG has a limit of 64 Initiator IQN's.  We have customers
with deployments >64 nodes, so this creates a problem.

To deal with this, we'll add the ability to configure multiple VAG's
(up to 4), each providing a mapping for 64 IQN's.  This means we can
increase the max node count from 64 to 256.

The way this works is that each VAG is a set of 64 IQN's (nodes), when
a volume is created from Trident the volume is added to each of the
specified VAGs.  Givent hat there is a limit of 4 VAG assignments per
volume that gets us to a max of 256 nodes.

Going forward we'll come up with better more dynamic ways to handle this
or perahps even switch to just using CHAP, but for now this should give
us a pretty good window to work with.

Needed by: Trident Github issue #24
Internal tracking ID: US2573