Detect and report Unauthorized requests on SF
Currently if a deployer fails to set up the base admin account
with the correct settings initialization of the driver will fail,
and this can be very difficult to identify.

This change adds a check in the API request response for any
unauthorized status responses and logs them, as well as returning
an error.  The result is that init catches the error and does a
fatal log message indicating that the request was unauthorized.

Addresses Github issue: #66