Add ability to specify QoS values on Clone
Currently when we do a clone, it's a clone, and that includes
all of the settings etc from the source volume (including QoS).
NDVP provides an option to specify new QoS settings though which
is great, but the SolidFire driver doesn't actually check the
options passed in on a Clone request, so those are just completely
ignored.

This patch adds the logic to the CloneVolume method to check if
QoS options were passed in, and if they were it will honor those
and set them on the destination volume.  In the future we might
want to determine if we need to do similar things for other options
like size etc.  Currently we don't seem to prohibit passing anything
in here, so we may need to either check that higher up and disallow
certain things, or probably need to implement any valid option.

Addresses GH Issue: #85