Set 512e to true by default, and allow option
SolidFire by default should always use 512e, unfortunately
we weren't doing this, so two things need to happen:
1.  Allow a config variable for default block size (512|4096)
    and default to 512 when not specified.
2.  Provide an option to configure it on create via
    ``-o blocksize=512|4096

This change adds a config entry ``DefaultBlockSize`` that
can be set to either 512 or 4096 which will be used when
no -o blocksize= options is specified on create.

Please note, this does not apply to things like Clone or Create
from Snapshot.

Additionally, changes to ndvp config structure (default section etc)
have merged that aren't documented in the SolidFire section of the docs
and haven't been implemented in a number of sections in the driver.
Will provide a follow up patch to try and clean that up and implement
a defaults section as well as make the json variables consistent with
the reset of the NDVP drivers.

Related to Trident Github Issue #33